Add a link to JackTrip
I haven't seen a mention of JackTrip on the list. When I needed to
create a open source based network setup, JackTrip was the most readily
available solution. There is a JackTrip application which provides a
bridge between network and the JACK daemon. While the current UI and
overall use is geared towards playing live with a musician in another
location, JackTrip also allows for low latency transfer of multichannel
audio over LAN - this was my use. Since I personally tested it out, and
it worked well, I think it might deserve a spot on this list. I'm happy
to provide more information if needed.